docs: add JSDocs to task/*, minor refactoring of service and repository
