+++ src/resources/tasks/task.memory.repository.js
@@ -1,13 +1,24 @@
 const path = require('path');
 
-const Task = require('./task.model');
 const { DATA_PATH } = require('../../common/config');
 const createGetDataFromFile = require('../../utils/create.get.data.from.file');
 const createSaveDataToFile = require('../../utils/create.save.data.to.file');
 
+/**
+ * Represents task memory repository error
+ * @class
+ */
 class TaskMemoryRepositoryError {
+  /**
+   * Creates task memory repository error
+   * @constructor
+   * @param {string} message - the message
+   */
   constructor(message = 'unknown error') {
+    /**  @member {number} */
     this.status = 503;
+
+    /** @member {string} */
     this.message = `Task memory repository error: ${message}`;
   }
 }
@@ -16,11 +27,23 @@ const fileName = path.resolve(DATA_PATH, 'tasks.json');
 const readTasks = createGetDataFromFile(fileName, TaskMemoryRepositoryError);
 const saveTasks = createSaveDataToFile(fileName);
 
+/**
+ * Throws task memory repository error
+ * @param {Error} e - the error object
+ * @param {string} message - the message
+ * @throws {TaskMemoryRepositoryError}
+ */
 const throwTaskRepositoryError = (e, message) => {
   process.stderr.write(e);
   throw new TaskMemoryRepositoryError(message);
 };
 
+/**
+ * Selects all tasks
+ * @async
+ * @returns {Promise<Task[]|[]|void>} - list of all tasks
+ * @throws {TaskMemoryRepositoryError}
+ */
 const getAll = async () => {
   try {
     return await readTasks();
@@ -29,13 +52,33 @@ const getAll = async () => {
   }
 };
 
-const getByBoardId = async boardId => (
-  (await getAll()).filter(({ boardId: taskBoardId }) => taskBoardId === boardId)
-);
+/**
+ * Selects tasks by board id
+ * @exports
+ * @async
+ * @param {string} boardId - The board id
+ * @returns {Promise<Task[]|[]|void>} - list of tasks for board
+ */
+const getByBoardId = async boardId => {
+  const tasks = await getAll();
 
+  return tasks.filter(task => task.boardId === boardId)
+};
+
+/**
+ * Returns task by id and board id
+ * @exports
+ * @async
+ * @param {string} taskId - the task id
+ * @param {string} boardId - the board id
+ * @returns {Promise<Task>}
+ */
 const getTask = async (taskId, boardId) => {
+  /** @type {Task[]} */
   const tasks = await getByBoardId(boardId);
-  const task = tasks.find(({ id }) => id === taskId);
+
+  const task = tasks.find(item => item.id === taskId);
+
   if (!task) {
     throw Object.create({ status: 404, message: 'Task not found' });
   }
@@ -43,57 +86,42 @@ const getTask = async (taskId, boardId) => {
   return task;
 };
 
-const create = async ({
-  title,
-  order,
-  description,
-  userId,
-  boardId,
-  columnId,
-}) => {
-  const tasks = await getAll();
-  const newTask = new Task({
-    title,
-    order,
-    description,
-    userId,
-    boardId,
-    columnId,
-  });
-  tasks.push(newTask);
+/**
+ * Creates new task
+ * @exports
+ * @async
+ * @param {Task} newTask - the new task
+ * @returns {Promise<boolean|void>} - flag of success
+ * @throws {TaskMemoryRepositoryError}
+ */
+const create = async newTask => {
   try {
+    const tasks = await getAll();
+    tasks.push(newTask);
     await saveTasks(tasks);
 
-    return newTask;
+    return true;
   } catch (e) {
     return throwTaskRepositoryError(e, 'task was not created');
   }
 };
 
-const update = async (task, {
-  title,
-  order,
-  description,
-  userId,
-  boardId,
-  columnId,
-}) => {
-  const updated = new Task({
-    ...task,
-    ...(title && { title }),
-    ...(order && { order }),
-    ...(description && { description }),
-    ...(userId !== undefined && { userId }),
-    ...(boardId !== undefined && { boardId }),
-    ...(columnId !== undefined && { columnId }),
-  });
+/**
+ * Updates tasks
+ * @exports
+ * @async
+ * @param {Task} updatedTask - The updated task
+ * @returns {Promise<boolean|void>} - the flag of success
+ * @throws {TaskMemoryRepositoryError}
+ */
+const update = async updatedTask => {
   try {
     const tasks = (await getAll()).map(cur => (
-      cur.id === task.id ? updated : cur
+      cur.id === updatedTask.id ? updatedTask : cur
     ));
     await saveTasks(tasks);
 
-    return updated;
+    return true;
   } catch (e) {
     return throwTaskRepositoryError(e, 'task was not updated');
   }
@@ -145,10 +173,18 @@ const clearTaskUserId = async userId => {
   }
 };
 
+/**
+ * Removes task by id
+ * @exports
+ * @async
+ * @param {string} taskId - the task id
+ * @returns {Promise<boolean|void>} - the flag of success
+ * @throws {TaskMemoryRepositoryError}
+ */
 const deleteTask = async taskId => {
-  let tasks = await getAll();
   try {
-    tasks = tasks.filter(({ id }) => id !== taskId);
+    let tasks = await getAll();
+    tasks = tasks.filter(task => task.id !== taskId);
     await saveTasks(tasks);
 
     return true;
